osutil: aggregate mockable symbols
This will help in upcoming improvements to mocking. As more things are
added to osutil it is sometimes unclear what is being mocked and how.
This patch should help to improve the situation.
